+++ ft_strncmp.c
@@ -6,17 +6,21 @@
 /*   By: marco-fe <marvin@student.42madrid.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/16 12:16:49 by marco-fe          #+#    #+#             */
-/*   Updated: 2023/03/17 11:09:12 by marco-fe         ###   ########.fr       */
+/*   Updated: 2023/03/30 12:33:51 by marco-fe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+/*Compares "n" bytes of "s1" and "s2". If finds a difference, returns
+ * "p_s1[i]" - "p_s2[i]" (p_ because of the pointers). Otherwise there are
+ * no difference, so returns 0.*/
+
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	unsigned char	*p_s1;
 	unsigned char	*p_s2;
-	unsigned int	i;
+	size_t			i;
 
 	p_s1 = (unsigned char *)s1;
 	p_s2 = (unsigned char *)s2;
@@ -32,9 +36,9 @@ int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
 /*
 int	main(void)
 {
-	char			text1[] = "Hola";
-	char			text2[] = "Holbs";
-	unsigned int	n = 3;
+	char	text1[] = "Hola";
+	char	text2[] = "Holbs";
+	size_t	n = 3;
 
 	printf("Real:\t%d\n", strncmp(text1, text2, n));
 	printf("Markoh:\t%d\n", ft_strncmp(text1, text2, n));
